// desktop/src/open-api/model/groupTaxRule.ts
/**
 * Receipt Wrangler API.
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface GroupTaxRule { 
    id: number;
    createdAt: string;
    createdBy?: number;
    /**
     * Created by entity\'s name
     */
    createdByString?: string;
    updatedAt?: string;
    /**
     * Group receipt settings foreign key
     */
    groupReceiptSettingsId?: number;
    /**
     * ISO 3166-1 alpha-2 country code the rule applies to
     */
    countryCode: string;
    /**
     * Whether tax from this country can be reclaimed
     */
    reclaimable?: boolean;
    /**
     * Whether a supplier tax id is required on the receipt to reclaim
     */
    requireTaxId?: boolean;
    /**
     * Fallback tax rate for this country
     */
    defaultRate?: number;
    /**
     * Optional human-readable label
     */
    label?: string;
}